@@ -2,9 +2,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from llmfoundry.models.layers.attention import (
-    ATTN_CLASS_REGISTRY, MultiheadAttention, MultiQueryAttention,
-    attn_bias_shape, build_alibi_bias, build_attn_bias, flash_attn_fn,
-    scaled_multihead_dot_product_attention, triton_flash_attn_fn)
+    ATTN_CLASS_REGISTRY, GroupedQueryAttention, MultiheadAttention,
+    MultiQueryAttention, attn_bias_shape, build_alibi_bias, build_attn_bias,
+    flash_attn_fn, scaled_multihead_dot_product_attention, triton_flash_attn_fn)
 from llmfoundry.models.layers.blocks import MPTBlock
 from llmfoundry.models.layers.custom_embedding import SharedEmbedding
 from llmfoundry.models.layers.fc import FC_CLASS_REGISTRY
@@ -17,6 +17,7 @@ __all__ = [
     'triton_flash_attn_fn',
     'MultiheadAttention',
     'MultiQueryAttention',
+    'GroupedQueryAttention',
     'attn_bias_shape',
     'build_attn_bias',
     'build_alibi_bias',
