@@ -12,7 +12,7 @@ try:
                                                log_config, pop_config,
                                                update_batch_size_info)
     from llmfoundry.utils.model_download_utils import (
-        download_from_cache_server, download_from_hf_hub)
+        download_from_hf_hub, download_from_http_fileserver)
 except ImportError as e:
     raise ImportError(
         'Please make sure to pip install . to get requirements for llm-foundry.'
@@ -28,7 +28,7 @@ __all__ = [
     'build_tokenizer',
     'calculate_batch_size_info',
     'convert_and_save_ft_weights',
-    'download_from_cache_server',
+    'download_from_http_fileserver',
     'download_from_hf_hub',
     'get_hf_tokenizer_from_composer_state_dict',
     'update_batch_size_info',
