@@ -6,9 +6,11 @@ import torch
 from omegaconf import OmegaConf as om
 
 from llmfoundry.models.layers import attention
-from llmfoundry.models.layers.attention import is_flash_v2_installed
+from llmfoundry.models.layers.attention import (check_alibi_support, gen_slopes,
+                                                is_flash_v2_installed)
 from llmfoundry.models.mpt.modeling_mpt import (apply_sequence_id,
                                                 gen_attention_mask_in_length,
+                                                gen_flash_attn_padding_info,
                                                 gen_rotary_embedding)
 
 
@@ -20,7 +22,7 @@ def allclose_helper(t0: torch.Tensor,
 
 
 @pytest.mark.gpu
-@pytest.mark.parametrize('attn_impl_0,attn_impl_1', [
+@pytest.mark.parametrize('attn_impl_0, attn_impl_1', [
     ('flash', 'triton'),
     ('flash', 'torch'),
     ('triton', 'torch'),
@@ -74,9 +76,9 @@ def test_attn_impl(attn_impl_0: str,
     """
     alibi = pos_emb_config['alibi']
     rope = pos_emb_config['rope']
-    if alibi and (attn_impl_0 == 'flash' or attn_impl_1 == 'flash'):
-        pytest.skip('flash attn does not support alibi')
-
+    if alibi and not (check_alibi_support(attn_impl_0) and
+                      check_alibi_support(attn_impl_1)):
+        pytest.skip('flash attention below v2.4.2 does not support alibi.')
     if rope and (pos_emb_config['rope_impl']
                  == 'dail') and (not is_flash_v2_installed()):
         pytest.skip('dail implementation of rope requires flash attention 2.')
@@ -163,12 +165,25 @@ def test_attn_impl(attn_impl_0: str,
         attn_uses_sequence_id=attn_uses_sequence_id,
         attn_impl=attn_impl_0,
         attention_mask=attention_mask)
+
+    flash_attn_padding_info_0 = {}
+    if attn_impl_0 == 'flash':
+        flash_attn_padding_info_0 = gen_flash_attn_padding_info(
+            n, s, 0, torch.device(device), attention_mask_in_length_0,
+            attention_mask)
+
     attention_mask_in_length_1 = gen_attention_mask_in_length(
         sequence_id=sequence_id,
         S=s,
         attn_uses_sequence_id=attn_uses_sequence_id,
         attn_impl=attn_impl_1,
         attention_mask=attention_mask)
+
+    flash_attn_padding_info_1 = {}
+    if attn_impl_1 == 'flash':
+        flash_attn_padding_info_1 = gen_flash_attn_padding_info(
+            n, s, 0, torch.device(device), attention_mask_in_length_1,
+            attention_mask)
 
     x0 = torch.randn(n, s, f).to(device)
     x1 = x0.clone().detach()
@@ -177,6 +192,12 @@ def test_attn_impl(attn_impl_0: str,
 
     with torch.autocast(x0.device.type):
         attn_bias_0 = gen_bias(attn_impl_0)
+        alibi_slopes_0 = None
+        if alibi and attn_impl_0 == 'flash':
+            alibi_slopes_0 = gen_slopes(n_heads=cfg.n_heads,
+                                        alibi_bias_max=8,
+                                        device=torch.device(device),
+                                        return_1d=True)
         rotary_emb_w_meta_info = None
         if rope:
             rotary_embedding = gen_rotary_embedding(
@@ -209,15 +230,23 @@ def test_attn_impl(attn_impl_0: str,
                          attention_mask=attention_mask,
                          rotary_emb_w_meta_info=rotary_emb_w_meta_info,
                          is_causal=True,
-                         attention_mask_in_length=attention_mask_in_length_0)
+                         flash_attn_padding_info=flash_attn_padding_info_0,
+                         alibi_slopes=alibi_slopes_0)
         attn_bias_1 = gen_bias(attn_impl_1)
+        alibi_slopes_1 = None
+        if alibi and attn_impl_1 == 'flash':
+            alibi_slopes_1 = gen_slopes(n_heads=cfg.n_heads,
+                                        alibi_bias_max=8,
+                                        device=torch.device(device),
+                                        return_1d=True)
         y1, _, _ = attn1(x1,
                          past_key_value=None,
                          attn_bias=attn_bias_1,
                          attention_mask=attention_mask,
                          rotary_emb_w_meta_info=rotary_emb_w_meta_info,
                          is_causal=True,
-                         attention_mask_in_length=attention_mask_in_length_1)
+                         flash_attn_padding_info=flash_attn_padding_info_1,
+                         alibi_slopes=alibi_slopes_1)
         y0 *= attention_mask.unsqueeze(-1)
         y1 *= attention_mask.unsqueeze(-1)
 
@@ -298,11 +327,16 @@ def test_vs_mha(attn_impl: str, device: str = 'cuda'):
     x1.requires_grad = True
 
     with torch.autocast(x0.device.type):
+        flash_attn_padding_info = None
+        if attn_impl == 'flash':
+            flash_attn_padding_info = gen_flash_attn_padding_info(
+                n, s, 0, torch.device(device), None, attention_mask)
         y0, _, _ = mmhsa(x0,
                          past_key_value=None,
                          attn_bias=None,
                          attention_mask=attention_mask,
-                         is_causal=True)
+                         is_causal=True,
+                         flash_attn_padding_info=flash_attn_padding_info)
         y1, _ = tmhsa(x1,
                       x1,
                       x1,
@@ -372,11 +406,16 @@ def test_grouped_attention_heads(attn_impl: str,
     x0.requires_grad = True
 
     with torch.autocast(x0.device.type):
+        flash_attn_padding_info = None
+        if attn_impl == 'flash':
+            flash_attn_padding_info = gen_flash_attn_padding_info(
+                n, s, 0, torch.device(device), None, attention_mask)
         y0, _, _ = mmhsa(x0,
                          past_key_value=None,
                          attn_bias=None,
                          attention_mask=attention_mask,
-                         is_causal=True)
+                         is_causal=True,
+                         flash_attn_padding_info=flash_attn_padding_info)
         y0 *= attention_mask.unsqueeze(-1)
 
         loss0 = y0.sum()
