@@ -5,7 +5,7 @@ from typing import Union
 
 import pytest
 from composer import Trainer
-from composer.utils import get_device, using_torch_2
+from composer.utils import get_device
 from omegaconf import OmegaConf as om
 from torch.distributed.algorithms._checkpoint.checkpoint_wrapper import \
     CheckpointWrapper
@@ -65,9 +65,6 @@ def test_fsdp_act_checkpoint(activation_checkpointing: bool,
          ]:
         module = trainer.state.model.model._fsdp_wrapped_module.transformer.blocks[
             0]._fsdp_wrapped_module
-        if not using_torch_2():
-            module = trainer.state.model.model._fsdp_wrapped_module.transformer.blocks[
-                0]._fsdp_wrapped_module._fpw_module
         assert isinstance(module, CheckpointWrapper)
     elif activation_checkpointing_target == [
             'grouped_query_attention'
