@@ -16,11 +16,9 @@ from transformers.utils import SAFE_WEIGHTS_INDEX_NAME, SAFE_WEIGHTS_NAME
 from transformers.utils import WEIGHTS_INDEX_NAME as PYTORCH_WEIGHTS_INDEX_NAME
 from transformers.utils import WEIGHTS_NAME as PYTORCH_WEIGHTS_NAME
 
-from llmfoundry.utils.model_download_utils import (DEFAULT_IGNORE_PATTERNS,
-                                                   PYTORCH_WEIGHTS_PATTERN,
-                                                   SAFE_WEIGHTS_PATTERN,
-                                                   download_from_cache_server,
-                                                   download_from_hf_hub)
+from llmfoundry.utils.model_download_utils import (
+    DEFAULT_IGNORE_PATTERNS, PYTORCH_WEIGHTS_PATTERN, SAFE_WEIGHTS_PATTERN,
+    download_from_hf_hub, download_from_http_fileserver)
 
 # ======================== download_from_hf_hub tests ========================
 
@@ -103,15 +101,17 @@ def test_download_from_hf_hub_weights_pref(mock_list_repo_files: MagicMock,
                                            repo_files: List[str],
                                            expected_ignore_patterns: List[str]):
     test_repo_id = 'test_repo_id'
+    save_dir = 'save_dir'
     mock_list_repo_files.return_value = repo_files
 
-    download_from_hf_hub(test_repo_id, prefer_safetensors=prefer_safetensors)
+    download_from_hf_hub(test_repo_id,
+                         save_dir=save_dir,
+                         prefer_safetensors=prefer_safetensors)
     mock_snapshot_download.assert_called_once_with(
         test_repo_id,
-        cache_dir=None,
+        local_dir=save_dir,
         ignore_patterns=expected_ignore_patterns,
-        token=None,
-    )
+        token=None)
 
 
 @mock.patch('huggingface_hub.snapshot_download')
@@ -121,10 +121,11 @@ def test_download_from_hf_hub_no_weights(
     mock_snapshot_download: MagicMock,
 ):
     test_repo_id = 'test_repo_id'
+    save_dir = 'save_dir'
     mock_list_repo_files.return_value = []
 
     with pytest.raises(ValueError):
-        download_from_hf_hub(test_repo_id)
+        download_from_hf_hub(test_repo_id, save_dir)
 
     mock_snapshot_download.assert_not_called()
 
@@ -148,12 +149,12 @@ def test_download_from_hf_hub_retry(
     mock_snapshot_download.side_effect = exception
 
     with pytest.raises((tenacity.RetryError, exception.__class__)):
-        download_from_hf_hub('test_repo_id')
+        download_from_hf_hub('test_repo_id', 'save_dir')
 
     assert mock_snapshot_download.call_count == expected_attempts
 
 
-# ======================== download_from_cache_server tests ========================
+# ======================== download_from_http_fileserver tests ========================
 
 ROOT_HTML = b"""
 <!DOCTYPE html>
@@ -182,51 +183,47 @@ SUBFOLDER_HTML = b"""
 @mock.patch.object(requests.Session, 'get')
 @mock.patch('os.makedirs')
 @mock.patch('builtins.open')
-def test_download_from_cache_server(mock_open: MagicMock,
-                                    mock_makedirs: MagicMock,
-                                    mock_get: MagicMock):
-    cache_url = 'https://cache.com/'
-    model_name = 'model'
-    formatted_model_name = 'models--model'
+def test_download_from_http_fileserver(mock_open: MagicMock,
+                                       mock_makedirs: MagicMock,
+                                       mock_get: MagicMock):
+    model_url = f'https://cache.com/models/model/'
     save_dir = 'save_dir/'
 
     mock_open.return_value = MagicMock()
 
     def _server_response(url: str, **kwargs: Dict[str, Any]):
-        if url == urljoin(cache_url, f'{formatted_model_name}/blobs/'):
+        if url == model_url:
             return MagicMock(status_code=HTTPStatus.OK, content=ROOT_HTML)
-        if url == urljoin(cache_url, f'{formatted_model_name}/blobs/file1'):
+        if url == urljoin(model_url, 'file1'):
             return MagicMock(status_code=HTTPStatus.OK)
-        elif url == urljoin(cache_url, f'{formatted_model_name}/blobs/folder/'):
+        elif url == urljoin(model_url, 'folder/'):
             return MagicMock(status_code=HTTPStatus.OK, content=SUBFOLDER_HTML)
-        elif url == urljoin(cache_url,
-                            f'{formatted_model_name}/blobs/folder/file2'):
+        elif url == urljoin(model_url, 'folder/file2'):
             return MagicMock(status_code=HTTPStatus.OK)
         else:
             return MagicMock(status_code=HTTPStatus.NOT_FOUND)
 
     mock_get.side_effect = _server_response
-    download_from_cache_server(model_name, cache_url, 'save_dir/')
+    download_from_http_fileserver(model_url, save_dir)
 
-    mock_open.assert_has_calls([
-        mock.call(os.path.join(save_dir, formatted_model_name, 'blobs/file1'),
-                  'wb'),
-        mock.call(
-            os.path.join(save_dir, formatted_model_name, 'blobs/folder/file2'),
-            'wb'),
-    ],
-                               any_order=True)
+    mock_open.assert_has_calls(
+        [
+            mock.call(os.path.join(save_dir, 'file1'), 'wb'),
+            mock.call(os.path.join(save_dir, 'folder/file2'), 'wb'),
+        ],
+        any_order=True,
+    )
 
 
 @mock.patch.object(requests.Session, 'get')
-def test_download_from_cache_server_unauthorized(mock_get: MagicMock):
-    cache_url = 'https://cache.com/'
+def test_download_from_http_fileserver_unauthorized(mock_get: MagicMock):
     model_name = 'model'
+    cache_url = f'https://cache.com/models--{model_name}/blobs/'
     save_dir = 'save_dir/'
 
     mock_get.return_value = MagicMock(status_code=HTTPStatus.UNAUTHORIZED)
     with pytest.raises(PermissionError):
-        download_from_cache_server(model_name, cache_url, save_dir)
+        download_from_http_fileserver(cache_url, save_dir)
 
 
 @pytest.mark.parametrize(['exception', 'expected_attempts'], [
@@ -236,7 +233,7 @@ def test_download_from_cache_server_unauthorized(mock_get: MagicMock):
 ])
 @mock.patch('tenacity.nap.time.sleep')
 @mock.patch('llmfoundry.utils.model_download_utils._recursive_download')
-def test_download_from_cache_server_retry(
+def test_download_from_http_fileserver_retry(
     mock_recursive_download: MagicMock,
     mock_sleep: MagicMock,  # so the retry wait doesn't actually wait
     exception: BaseException,
@@ -245,4 +242,4 @@ def test_download_from_cache_server_retry(
     mock_recursive_download.side_effect = exception
 
     with pytest.raises((tenacity.RetryError, exception.__class__)):
-        download_from_cache_server('model', 'cache_url', 'save_dir')
+        download_from_http_fileserver('cache_url', 'save_dir')
